+++ tests/testthat/test-get_mcmc_samples.R
@@ -4,7 +4,7 @@ test_that("Default Bayesian hierarchical model can be build", {
   test <- get_mcmc_samples(data = dat, id_application = "application",
                            id_voter = "voter",
                            grade_variable = "num_grade",
-                           n_chains = 2, n_iter = 1000,
+                           n_chains = 2, n_adapt = 1000, n_iter = 1000,
                            n_burnin = 1000)
-  expect_equal(class(test), "mcmc")
+  expect_equal(class(test$samples), "mcmc")
 })